+++ src/source/Yaaf.DependencyInjection.SimpleInjector/SimpleInjectorBinder.cs
@@ -3,17 +3,19 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 
 namespace Yaaf.DependencyInjection.SimpleInjector
 {
-    public class SimpleInjectorBinder<TService> : IBinder<TService> where TService : class {
+    internal class SimpleInjectorBinder<TService> : IBinder<TService> where TService : class {
 
-        private Container container;
+        private SimpleInjectorKernel kernel;
         private Type serviceType;
         private bool allowOverride;
-        public SimpleInjectorBinder(Container container, Type serviceType = null, bool allowOverride = false) { 
-            this.container = container;
+        internal SimpleInjectorBinder(SimpleInjectorKernel kernel, Type serviceType = null, bool allowOverride = false)
+        {
+            this.kernel = kernel;
             this.serviceType = serviceType;
             this.allowOverride = allowOverride;
         }
@@ -30,49 +32,69 @@ namespace Yaaf.DependencyInjection.SimpleInjector
                 this.doOnEnd();
             }
         }
-        private IDisposable Begin()
+
+        internal Container Container { get { return kernel.Container; } set { kernel.Container = value; } }
+
+        private void Secure(Action doSomething)
         {
-            var before = container.Options.AllowOverridingRegistrations;
-            container.Options.AllowOverridingRegistrations = allowOverride;
-            return new EndHelper(() =>
-                container.Options.AllowOverridingRegistrations = before);
+            bool lockWasTaken = false;
+            var temp = kernel.LockObj;
+            var before = Container.Options.AllowOverridingRegistrations;
+            try {
+                Monitor.Enter(temp, ref lockWasTaken); 
+                {
+                    if (kernel.IsLocked)
+                    {
+                        Container = SimpleInjectorKernelCreator.Clone(Container);
+                        kernel.IsLocked = false;
+                    }
+                    Container.Options.AllowOverridingRegistrations = allowOverride;
+                    System.Diagnostics.Debug.Assert(!kernel.IsLocked);
+                    doSomething(); 
+                } 
+            }
+            finally {
+                if (lockWasTaken) Monitor.Exit(temp);
+                Container.Options.AllowOverridingRegistrations = before;
+            }
         }
+
         public void To(Type instanceType)
         {
-            using (this.Begin())
+            this.Secure(() =>
             {
-                container.Register(serviceType ?? typeof(TService), instanceType);
-            }
+                Container.Register(serviceType ?? typeof(TService), instanceType);
+            });
         }
 
         public void ToConstant<TInstance>(TInstance serviceInstance) where TInstance : TService
         {
-            using (this.Begin())
+            this.Secure(() =>
             {
                 if (serviceType == null)
                 {
-                    container.RegisterSingle<TService>(serviceInstance);
+                    Container.RegisterSingle<TService>(serviceInstance);
                 }
                 else
                 {
-                    container.RegisterSingle(serviceType, serviceInstance);
+                    Container.RegisterSingle(serviceType, serviceInstance);
                 }
-            }
+            });
         }
 
         public void To<TInstance>() where TInstance : class, TService
         {
-            using (this.Begin())
+            this.Secure(() =>
             {
                 if (serviceType == null)
                 {
-                    container.Register<TService, TInstance>();
+                    Container.Register<TService, TInstance>();
                 }
                 else
                 {
-                    container.Register(serviceType, typeof(TInstance));
+                    Container.Register(serviceType, typeof(TInstance));
                 }
-            }
+            });
         }
     }
 }