+++ doc/ReleaseNotes.md
@@ -1,4 +1,9 @@
-﻿### 1.1.0
+﻿### 1.1.0 (SimpleInjector 1.1.1)
+
+ * SimpleInjector implementation now supports changing services (workaround container locking)
+ * SimpleInjectpr now supports Rebind is now supported
+
+### 1.1.0
 
  * Added SimpleInjector support.